Image-Gen-Server/requirements.txt
# Core dependencies for minimal ComfyUI workflow
torch>=2.0.0
torchvision
torchaudio
numpy>=1.25.0
einops
transformers>=4.37.2
tokenizers>=0.13.3
sentencepiece
safetensors>=0.4.2
Pillow
scipy
tqdm
psutil
pyyaml

# GGUF support
gguf>=0.10.0

# Flask server
flask
flask-cors

# Optional but recommended
accelerate

# Additional ComfyUI dependencies
torchsde
opencv-python
imageio
imageio-ffmpeg